Replace the StyledNode trait with inherent methods.
